First Simple Module Implemented
One small snippet of the original code in core.js was cut and pasted into a simple component which is now imported in core.js. The two objects moved are the cache (which stores temporary data) and pressed object (which tracks key pressed). Ths is essentially a proof of concept for modularity as lots more is yet to be modularized.