Created SetUp module & modified Cache module
The SetUp.js module was created to include all the related functionality for setting up the document. Window blur event for resetting keys pressed was moved to the Cache.js module.

Minor changes include having the tools include the hide class in the HTML to be hidden by default. Other changes include renaming a variable from 'toggle' to 'settings' in SetUp, and adding clarification comments.