Rudimentary OOP implementation by drawing a Line.
Realizing possible inefficiencies with this method (especially since the HTML elements on the Web Page can be considered a form of Objects themselves, although not in the traditional sense).

Attempting to represent Elements using Java Syntax since there is a clear inheritance "is-a" relationship between an Element and a Subset of Element. However, there may be caveats to using this Java-like approach, as the Objects are self-encapsulated, which would imply that there should be a JS Object stored in memory for every HTML element that exists on the Web Page (inefficient).

Currently the app relies on creating, directly reading from and setting attributes directly to elements on the SVG Canvas. Is it really worth changing that, or changing it partially? An insight into this can be found on line 46 of the newSVG module, which will replace the Draw module, and a portion of the SVG class related to element creation.

Created Events module which contains the user events which trigger element creation or modify (move/resize them).

Note: JavaScript does not seem to support the Java convention of naming constructors the same as the Class name, instead the constructor has to be named constructor(), as seen in the Line object.

Removed Files that are not currently being used in Main directory to clear up headspace.