Removed Two Common Errors
One error was related to element selection. The check was for if the $(ele) passed in for the selection area was undefined. Although the variable was defined, the actual element that was being searched for by jQuery might not have existed. Used the length property in order to determine if a matching element was returned.

Another error involved the issue of passing in a jQuery reference $("") into another jQuery reference string expression when trying to retrieve a layers item by ID, as well as when passing the cache.ele object into the area function.