Simplifying the Selection Method
By using the already existing createAttr() used when creating elements, the bounds are easily redefined for existing elements without too much extra hastle. A little extra logic for the Line Element though. Commented out functions in Resize with intent to remove soon.

There is still a bug when dragging the left and top handle to the opposite edge of the object, where it doesn't stay put.

Another change was removing an offset in the UI for the selection.