TextBox Event made functional (needs improving)
The mouseup event creates a textbox area, but the user is still unable to input data to the textbox.

A duplicate element gets created on the on the SVG Canvas (which will actually get exported). The Textbox being visibly created is just a placeholder (a duplicate has to be updated in the #editor behind the scenes).