Adding Text officially supported, but not moving
Added extra check for Cache delete to make sure an input is not currently focused (which would indicate user is editing text). Prevent delete in that case.

Added a check for editSVG Object. Without the check that the element wasn't of type `text`, the large thick default stroke border would have been applied to it (making text unreadable).