Layer Tab UI Functional
It now includes multiple layer outlining/highlighting with shift+click, smooth dragging/reorder, and element selection is synchronized between the layers tab and canvas.

Although the UI interaction is fully functional for the layers tab, that does not hold consistent for interactions between the layers tab and other contexts during multiple element selection.

Fixed Tab Dragging for all tabs.

Fixed drawing when there is a canvas offset but only when making window larger, but not for zooming or dragging elements yet.