Zoom Fixes
Moving and creating SVG elements while zoomed is fully functional. Resizing elements mostly works except with shift and cmd keys ocassionally (with lines).

Using getBBox() in svg.initial ensures correct behavior for SVG elements and makes modifications easier. It returns values relative to the SVG container (editor) and does NOT account for transformations.

In contrast, getBoundingClientRect() returns relative to the viewport and accounts for transformations. This is used for adjusting the cursor position based on the zoom and/or canvas margins.

previewMove() is also updated to account for these changes.

Element modifications do not yet properly account for the scale transformation.