Color module improvements
This includes several bug fixes and tweaks to improve the functionality and stability of the color module. Right clicking schemes (leftmost section) shows the color previews for the different scheme collection options (HTML, Web Safe, Flat Design, etc.).

Color saving now works correctly whether right click to add to favorites or double clicking to save to the tuner.

When the color search field is emptied, then the previous scroll position is restored for any tab.

The color picker is automatically updated with the colors from the canvas when a file is loaded and when new element colors are added.