Scale transform Resizing support
Pressing 'ctrl' while resizing/moving an element will apply a scale/translate transformation on that element. Group elements are scaled by default, even when 'ctrl' is not pressed. 'ctrl+shift' and 'ctrl+cmd' not yet supported.

type === 'genericElement' indicates we are working with generally-applicable transforms versus an element's own attributes. As such, <g> or unsupported elements are treated as a genericElement. This is what enables the user to scale or translate an element with 'ctrl'.

Allow creating a new nested group when dragging an existing group over a non-group element layer.

Changed previewMove implementation to easily handle any element translation (when pressing shift or cmd), and now the previewMove drag line and number are visible outside the bounds of the SVG canvas (through a separate SVG overlay).

newSVG is updated to simply insert the new element in front of the current element (cache.ele) regardless of whether it is a group or not.