Import and image fixes, and other improvements
Prevent caching of the application, in order to ensure the latest version is always displayed.

Restrict imports on the start page to just '.svg' files. Widen support for image files when importing into an SVG canvas (by adding `image/*` to the accept attribute on the input tag).

Prevent zooming the canvas when the user is on the start page (or if the canvas is not displayed).