Tab re-orientation fix and fix layer tab dragging
Added noReorient class to tabs that do not support vertical/horizontal re-orientation to prevent glitchy behavior after pushing/dragging out when tab drag handle is snapped to the screen edge.

When a specific layer was clicked, e.stopPropagation() would stop the layers panel from being moved to the front. Similarly, if the layer panel was being dragged when it was last clicked, then that caused stopPropagation() to trigger when the user incidentally mouses over the listed layers.

Included latest build version, as it was not included in the previous commit.