Crisp Zooming Supported
Previously the scale transform was used which caused a loss of SVG quality while zooming in. Now the width and height values are modified instead, which allow the quality of the SVG to be preserved while scaling up.

Will keep previous option of scale zooming in case it yields performance advantages in some scenarios (with option to toggle).

Need to center the zooming to a smaller area, instead of mouse position being relative to the entire canvas.