Rename app.html to index.html, and update vite config.
Since the new workflow now serves GitHub Pages directly from the dist/ folder, the app.html file was renamed to index.html so that GitHub pages will detect it automatically in dist/ directory for the root project path '/'.

Link in the readme was also updated to be valid relative to the dist/ directory.