TypeScript Checking Added as Vite plugin
TypeScript checking is desired for the new React files and logic. As the code is potentially refactored, then it will be gradually moved into the svgem-react directory, so new code should include proper linting.

For this reason, checkJS is set to false in svgm-react tsconfig in order to exclude files which are outside the svgem-react directory (non-jsx/tsx files).